torus works (limited paramaters, R>r, R & r < 10, distance torus not too far)

+++ src/plane.c
@@ -35,8 +35,8 @@ double	get_pl_angle(t_scene *scene, int num[2], t_vect3d Phit, t_vect3d *N)
 	return (angle);
 }
 
-//finds smallest non-negative t_value
-double	find_smallest(t_scene *scene, double *t, int num, int amount)
+//finds smallest non-negative t-value and returns which t_value this is
+int	find_smallest(t_scene *scene, double *t, int num, int amount)
 {
 	int	i;
 	int	j;
